+++ caching-service/src/main/java/org/zowe/apiml/caching/CachingService.java
@@ -17,13 +17,14 @@ import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.retry.annotation.EnableRetry;
 //import org.zowe.apiml.enable.EnableApiDiscovery;
+import org.zowe.apiml.enable.EnableApiDiscovery;
 import org.zowe.apiml.product.logging.annotations.EnableApimlLogger;
 import org.zowe.apiml.product.service.ServiceStartupEventHandler;
 
 import jakarta.annotation.Nonnull;
 
 @SpringBootApplication
-//@EnableApiDiscovery TODO fix for microservice
+@EnableApiDiscovery //TODO fix for microservice
 @EnableRetry
 @EnableApimlLogger
 public class CachingService implements ApplicationListener<ApplicationReadyEvent> {
