+++ caching-service/src/main/java/org/zowe/apiml/caching/service/inmemory/InMemoryStorage.java
@@ -13,7 +13,7 @@ package org.zowe.apiml.caching.service.inmemory;
 import lombok.extern.slf4j.Slf4j;
 import org.zowe.apiml.cache.StorageException;
 import org.zowe.apiml.cache.Storage;
-import org.zowe.apiml.cache.KeyValue;
+import org.zowe.apiml.caching.model.KeyValue;
 import org.zowe.apiml.caching.service.*;
 import org.zowe.apiml.caching.service.inmemory.config.InMemoryConfig;
 import org.zowe.apiml.message.core.MessageService;
