+++ caching-service/src/main/java/org/zowe/apiml/caching/service/vsam/VsamKey.java
@@ -12,7 +12,7 @@ package org.zowe.apiml.caching.service.vsam;
 
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
-import org.zowe.apiml.cache.KeyValue;
+import org.zowe.apiml.caching.model.KeyValue;
 import org.zowe.apiml.caching.service.vsam.config.VsamConfig;
 
 import java.io.UnsupportedEncodingException;
