+++ caching-service/src/main/java/org/zowe/apiml/caching/service/vsam/VsamStorage.java
@@ -12,7 +12,7 @@ package org.zowe.apiml.caching.service.vsam;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.retry.annotation.Retryable;
-import org.zowe.apiml.cache.KeyValue;
+import org.zowe.apiml.caching.model.KeyValue;
 import org.zowe.apiml.caching.service.EvictionStrategy;
 import org.zowe.apiml.caching.service.Messages;
 import org.zowe.apiml.cache.Storage;
