+++ caching-service/src/test/java/org/zowe/apiml/caching/service/redis/RedisStorageTest.java
@@ -13,7 +13,7 @@ package org.zowe.apiml.caching.service.redis;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.zowe.apiml.cache.KeyValue;
+import org.zowe.apiml.caching.model.KeyValue;
 import org.zowe.apiml.caching.service.Messages;
 import org.zowe.apiml.cache.StorageException;
 import org.zowe.apiml.caching.service.redis.exceptions.RedisOutOfMemoryException;
