+++ caching-service/src/test/java/org/zowe/apiml/caching/service/vsam/VsamKeyTest.java
@@ -12,7 +12,7 @@ package org.zowe.apiml.caching.service.vsam;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.zowe.apiml.cache.KeyValue;
+import org.zowe.apiml.caching.model.KeyValue;
 import org.zowe.apiml.caching.service.vsam.config.VsamConfig;
 
 import static org.hamcrest.MatcherAssert.assertThat;
