+++ integration-tests/src/test/java/org/zowe/apiml/integration/authentication/pat/AccessTokenServiceTest.java
@@ -69,7 +69,7 @@ public class AccessTokenServiceTest {
                 .body(bodyContent)
             .when()
                 .delete(REVOKE_ENDPOINT)
-            .then()
+            .then().log().ifValidationFails()
                 .statusCode(204);
             IntStream.range(0, 3).forEach(x -> {
                 given()
